requirements.txt

Django==3.2
openai==0.28.1
# pandas==1.5.2
python-dotenv==1.0.1
gunicorn==19.7.1


